Switched back to the nrf_ functions
From Eric:
I made some progress on this front. Basically, I think it worked before because we weren't using modular firmware yet. Once we moved to that, then the calls to the HAL_ functions inside the pinModeFast functions were done through the modular table, which isn't so fast. Before we were able to call the nrf_ functions directly which are much faster since they are inline.

So, I made a change to platform_headers.h, which allows us to bubble up commands from the system world to the user world. This means we can call the nrf_ functions directly again.